Improve measure! accuracy
Autobody measure is now forced to return Float64 for distance, which 
forces 0≤μ₀≤1. measure! also scans a slightly bigger region near the 
body surface to fill V. These are important when calculating σᵥ to 
ensure solver! converges.

+++ src/AutoBody.jl
@@ -31,7 +31,7 @@ function measure(body::AutoBody,x::Vector,t::Real)
     # Use DiffResults to get Hessian/gradient/value in one shot
     result = DiffResults.HessianResult(x)
     result = ForwardDiff.hessian!(result, x -> body.sdf(x,t), x)
-    d = DiffResults.value(result)
+    d = Float64(DiffResults.value(result))
     n̂ = DiffResults.gradient(result)
     κ = curvature(DiffResults.hessian(result))
